requirements.txt
annotated-types==0.7.0
appdirs==1.4.4
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Flask==2.0.2
Flask-Cors==4.0.0
frozendict==2.4.4
gevent==23.9.1
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.6.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==21.0.1
html5lib==1.1
httplib2==0.22.0
idna==3.3
importlib-metadata==4.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.9.1
marko==2.0.3
MarkupSafe==2.0.1
multitasking==0.0.10
numpy==1.26.4
packaging==24.0
pandas==2.2.2
peewee==3.17.5
plotly==5.22.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1-modules==0.4.0
pydantic==2.7.3
pydantic-core==2.18.4
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.5
requests==2.31.0
rsa==4.9
six==1.16.0
soupsieve==2.5
tenacity==8.5.0
tqdm==4.66.4
typing-extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.8
webencodings==0.5.1
Werkzeug==2.0.2
yfinance==0.2.37
zipp==3.7.0
zope.event==5.0
zope.interface==6.4.post2
